Added block placement and naming
You can now place and name the block, you will be prompted asking for a name once placed and when named it will be added to the map. I have had to make the wait a map so that I could get the location of the block. Currently it will not allow multiple tac inserts but that can be changed.